test: split benchmark cases and procedures
